moved scanExtensions to its own file
